QC08-004 crash on a generic parent with generic child corner case
The corner case covered by acats ca11021 is currently not supported; it
appears to require a very subtle navigation in the AST to find the
interesting entity (or perhaps the entity is not even in the AST, in which
case we need to adapt the connection rules with some magic).

* gnat2why/flow/flow_visibility.adb
(Make_Info): crash with explicit message to mark unsupported construct.

Change-Id: I747d623e89d2de976621738f0dcf222af8ff50e0