Q508-010 avoid duplicates in direct calls to nested packages
For a code with doubly-nested packages, e.g.

   procedure Proc is
      package Pkg1 ...
         package Pkg2 is ...

we were getting Pkg2 as a direct callee of both Pkg1 and Proc. Now we only
get Pkg2 as a direct callee of Pkg1, and Pkg1 as a direct callee of Proc.
This is simpler and better captures the idea of "directness".

This commit doesn't affect any outputs, but prepares the code for a proper
generation of the Global contract where the global is referenced by a nested
package.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Package_Declaration): only pick immediately nested packages.

Change-Id: I025014a42a4f1eb58419a899e51ef6a6220c165a