Q508-010 elaboration of nested packages contribute into Global
Nested packages with Initializes contract have their reads and writes are
already inlined in the CFG (this was already done); those without the that
contract need to be processed by the GG just like subprogram calls (this is
what this patch fixes).

* gnat2why/flow/flow-slice.adb
(Compute_Globals): pick elaboration of nested packages without Initializes
 contract.

* gnat2why/flow/flow_generated_globals-partial.adb
(Is_Callee): adapt so that it returns True for nested, library-level
 packages. Needed because this routine is used by the GG as a predicate for
 Callee_Set that represents callees of an analysed entity.

* testsuite/gnatprove/tests/Q508-010__flow_nested/
Enable and add a correct output (primary test for this TN).

* testsuite/gnatprove/tests/P929-043__flow_pkg_elaboration_in_direct_calls/
The output is not yet correct, because of other issues; the entity A is now
correctly recognized as an In_Out during GG, but then pruned because of how
non-terminating routines are handled. Accept for now.

Change-Id: I22ef39039c2b02aea4e305aed57aa1626f7359c4