Q810-010 adapt testsuite after frontend fix
With a frontend fix related to Part_Of and generic instances.

* testsuite/gnatprove/tests/QA04-034__volatile/test.out
Add missing Part_Ofs; this was a big reproducer related to volatile effects
in functions, and we want to keep it as a regression test for that.

* testsuite/gnatprove/tests/Q810-010__part_of_constituent/test.out
The TN was about improving the original message, which I think is exactly
what happened.

Change-Id: I94e75bccfd46764bd6fd0743a8c0034cd5760cee