Updated git submodules
Project: why3 master 88275407ff999b9b6d52282158c38c5bd5f5104a

R516-007 Locate why3server using current executable location

This commit changes the way why3server is executed in gnatwhy3 debug mode
and in gnat_server. It is not assumed to be in the PATH anymore. It now
uses the fact that location of gnatwhy3, gnat_server and why3server is
always the same (for customers and developpers).

* src/driver/prove_client.ml
(connect_internal): Localize why3server in the same directory as the one
of the executing file (gnatwhy3 or gnat_server).

Change-Id: I8dddd46990ba71d7f8b11e3fbb29fd1e1b204c4a