R622-041 assume pre of enclosing subprogram in nested packages
Assume precondition of the enclosing subprogram when verifying
elaboration of nested packages. We only do that for packages
declared in the declarative part of a subprogram.
Also assume dynamic property of in parameter of enclosing
subprograms.

* gnat2why/spark/spark_util.adb
(Directly_Enclosing_Subprogram_Or_Entry): Returns the entity of the
 first subprogram or entry enclosing E. Returns Empty if there is
 no such subprogram or if something else than a package (a
 concurrent type, a block statement...) is encountered in between.

* gnat2why/why/gnat2why-expr.adb
(Assume_Value_Of_Constants): Also supports in parameters if they
 are declared in enclosing scopes.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_VCs_For_Package_Elaboration): Assume precondition of the
 enclosing subprogram for nested packages which are declared
 directly inside the declarative part of a subprogram.

* tests/QC11-014__flow_initializes/test.out
Update test output.

Change-Id: Id7caab85e4cf5943455a430c35b3f24f3a3f525b