R709-026 enable ckecks and use ada equality in membership tests
The frontend transforms membership tests with a single alternative
of an object kind into an equality. When there are more than one
alternative, we were missing checks on alternatives of an object
kind and wrongly using why equality.

* gnat2why/why/gnat2why-expr.adb
(Transform_Alternative): Transform the expression in the
 alternative using the correct domain and use Ada equality.

* gnat2why/why/why-gen-expr.adb
(New_Ada_Equality): Also use user defined equality on limited
 types.

* tests/R709-026__membership
New test for membership tests.

Change-Id: Ia904941eacd6782e7e056a689576c6060f5cf76b