K516-007 Fix after detection of inconsistent postcondition
The warning for inconsistent postcondition is based on the generation of
Why3 code that may need to read the value returned by the function.
Add the declaration for the result object even when the function body is
not in SPARK for that case.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_VCs_For_Subprogram): Always declare result object.

Change-Id: I3d62d27db7f056a0d102cb1b03384d572228d717