R409-005 avoid unreliable calls to Parent in aliasing check
When checking a call that has been inlined for proof for aliasing, we are
analysing the original AST, where calls to Parent are no longer reliable.

As a workaround, we count how deep we down/up-traversed the AST.

* gnat2why/flow/flow-analysis-antialiasing.adb
(Find_Root): convert to procedure; add Depth parameter.
(Up_Ignoring_Conversions): same as above.

* testsuite/gnatprove/tests/R409-005__inline_aliasing_acats/test.out
Enable; add output.

Change-Id: I6815fc628c0195fef0810bec13177680acf9f5f6