explicitly ignore irrelevant nodes in Get_Variables
When Get_Variables is called not for expression (which is its canonical use)
but for a statement (when called from First_Variable_Use), we want to not
descent into certain constructs, e.g. entry bodies.

Curiously enough, this patch doesn't cause any diffs, but this is most
likely because the affected constructs are seldom used. Still, it seems
right to ignoring them.

(no-tn-check)

* gnat2why/flow/flow_utility.adb
(Get_Variables): do not traverse into irrelevant nodes.

Change-Id: I606e7f4ed98b385bc80f3759b2cc1ccde207aeeb