QC08-004 fix Get_Enclosing_Flow_Scope for private child units
For a hierarchy like in the Q908-011__flow_child_of_private_child test:

   package Foo ...
     private package Foo.PC
       package Foo.PC.C

when calling Get_Enclosing_Flow_Scope on Foo.PC.C we are interested on the
"private" keyword being attached to the package.

With this patch we will pick exactly that keyword; the previous code, with
Is_Private_Descendant, was effectively picking the "private" keyword from
any of the parents, which was not what we wanted.

There is no dedicated test provided, because the problem is difficult to
reproduce. This will come under the umbrella of fixing the visibility
issues; in fact, this is inspired by a similar check in
Flow_Visibility.Make_Info.

* gnat2why/flow/flow_refinement.ads
(Get_Enclosing_Flow_Scope): replace Is_Private_Descendant with
 Private_Present.

Change-Id: I889cc3b22b9670d0e4737ea87d7a91ddd29f33db