PA03-014 fix test to make it crash again
This test originally crashed due to a LIMITED WITH clause; then we
wishlisted and XFAIL-ed it. Now it doesn't crash, because the frontend
terminates analyses on an illegal Refined_Post.

This patch removes the illegal contract from the test, so that it crashes
again, which is what it is meant to demonstrate.

* testsuite/gnatprove/tests/PA03-014__limited_with/aida_character.adb
Remove illegal Refined_Post contract.

Change-Id: I8abc4bafdc81f87a21f062af07966ec1171357d5