PA03-014 fix crash while emitting an error message
In marking we detect types coming from LIMITED WITH as unsupported and crash
(at least in debug builds) on access to an uninitialized data while emitting
the error message. This happens when the Errout machinery tries to
substitute the "&" special character in the formal string passed to
Mark_Unsupported and trips on no data being provided for this substitution.

Now fixed; we get a nice error message that is nicely propagated to the
where the incomplete type is used.

* testsuite/gnatprove/tests/PA03-014__limited_with/
Enable (as there is no need to keep it dead); update output.

Change-Id: I321f308bf04a4d90396d4c901363f11fc0c0fca8