R706-020 add two lemmas about monotonicity of exponentiation
Add lemmas to the spark lemma library stating that the result of the
exponentiation increases when either the value or the exponent increases.

* include/spark-arithmetic_lemmas.ads
Add new lemmas about monotonicity of exponentiation.

* manual_proof.in
The new lemmas are proved manually.

* proof/Coq/common/lemma_exp_monotonic.prf
* proof/Coq/common/lemma_exp_monotonic_2.prf
Proof scripts for lemmas about exponentiation.

* proof/Coq/spark-integer_arithmetic_lemmas
* proof/Coq/spark-long_integer_arithmetic_lemmas
Add Coq VCs for the new lemmas using the proof scripts.

* proof/sessions
Update sessions.

Change-Id: I0d56e907d8e6ccd94987d620e8f412f0fff18d94