Q927-016 use proper model_trace for discriminants
We did not supply projection function for discriminants of record types.
This resulted in missing values of discriminants in some cases, as well as
in discriminant values printed as the value of the whole record type.
Also add assertions to check that the counterexamples have the expected
form. They should hold after this fix.

* gnat2why/why/gnat2why-counter_examples.adb
(Refine_Aux): Add assertions to check that the counterexamples have the
 expected form.

* gnat2why/why/why-gen-records.adb
(Declare_Record_Type): Also emit projection functions for discriminants.

* tests/K505-010__stack/test.out
* tests/K624-029__amortize_queue/test.out
* tests/M611-052__discr/test.out
* tests/M613-039__expr_actions/test.out
* tests/O512-022__stacks/test.out
* tests/O901-014__formal_vectors/test.out
* tests/Q529-037__discriminant/test.out
* tests/R618-026__hidden_comp_in_counterex/test.out
Improved counterexamples.

Change-Id: Ib881a81b85c05feb64f0574783c7fa1846e654e2