R717-007 Print fixed as Integers when possible in counterexamples.
When the exact result of the multiplication of the Small of a fixed type
and the integer returned as counterexample for this fixed type is an
integer, we print the fixed point number as integers.

* gnat2why/why/ce_pretty_printing.ad*
(Print_Fixed): Take an integer and the Small of a fixed point type and
return a pretty printed fixed point number.

* gnat2why/why/gnat2why-counter_examples.adb
(Print_Float): Remove comment
(Refine_Aux): Change fixed point case to a call to Print_Fixed.

* testsuite/gnatprove/tests/KA11-014__fixed_points/test.out
* testsuite/gnatprove/tests/O708-018__root_conversion/test.out
* testsuite/gnatprove/tests/O813-018__fixed_points/test.out
* testsuite/gnatprove/tests/P615-019__silver/test.out
* testsuite/gnatprove/tests/Q906-011__counterexample/test.out
* testsuite/gnatprove/tests/R220-054__fixed_point_small/test.out
* testsuite/gnatprove/tests/R722-002__manual/test.out

Update output

Change-Id: I3133e5134d7446c0cfde88397ee3fd8f49112f9b