R719-008 Restore counterexamples for z3
Last version of Z3 changes the way they print counterexamples. This commit
adds an option to the gnatprove.conf so that the config used for z3_ce
(which is z3 used for counterexample) has the option model_compress=false
which restores the default output of counterexamples.

* share/spark/config/gnatprove.conf
(z3_ce): add option

* testsuite/gnatprove/tests/QA25-058_z3_riposte_arrays/test.out
Restore output
* testsuite/gnatprove/tests/QA25-058_z3_riposte_arrays/test.opt
Remove file

Change-Id: I07210c626cec1d1e50faa63c4a571183d21e48ae