R727-014 Fix style checks on lemma library
SPARK lemma library is compiled with -gnatygo-u and should not lead to
style violations with this switch. Fix existing violations and add a test
to ensure any future violations will be detected.

* include/spark-test_array_lemmas.adb
Fix alphabetical order and casing.

* include/spark-unconstrained_array_lemmas.adb
Remove "in" for parameter.

* testsuite/gnatprove/tests/R727-014__lemma_library/test.gpr
* testsuite/gnatprove/tests/R727-014__lemma_library/test.py
* testsuite/gnatprove/tests/R727-014__lemma_library/test_lemmas.adb
Add test that compiles all the lemma library with -U, then generates
an executable using a few lemmas and run it.

Change-Id: I8c547494ff17586b187ad62cf7773dc37dc4727f