R629-018 Issue warning with -u when main unit not marked with SPARK_Mode
When calling GNATprove with switch -u to analyse only some units (typically
a single unit in particular inside the IDE with Examine/Prive File), it can
be surprising for users that the body is not analyzed when SPARK_Mode is not
applied directly or indirectly to that body.

The generic warning that no bodies were analyzed is not enough, as it only
applies when no results of analysis are present, and would not trigger also
if the spec is in SPARK.

So we now issue a specific warning when the main unit whose analysis is
requested is not marked with SPARK_Mode On or Off. This only applies for
modes flow and proof, not check.

* common/gnat2why_args.adb
* common/gnat2why_args.ads
Add variable Limit_Units for gnatprove to give the information to gnat2why
that switch -u was used.

* gnat2why/why/gnat2why-driver.adb
(GNAT_To_Why): Issue a warning when doing flow analysis or proof but the
 main unit is not in SPARK.

* gnatprove/gnatprove.adb
(Pass_Extra_Options_To_Gnat2why): Add the new component.

* testsuite/gnatprove/tests/R629-018__warning_no_analysis/pack.adb
* testsuite/gnatprove/tests/R629-018__warning_no_analysis/pack.ads
* testsuite/gnatprove/tests/R629-018__warning_no_analysis/test.gpr
* testsuite/gnatprove/tests/R629-018__warning_no_analysis/test.out
* testsuite/gnatprove/tests/R629-018__warning_no_analysis/test.py
Add test.

* testsuite/gnatprove/tests/MB13-068__no_analysis/test.out
Update test output.

Change-Id: Iac2ffeb1d0d24687bc473927a605d353f12782f1