QB20-019 Add check for global Inputs that should be Proof_Ins
In flow we were missing a check related to SPARK RM 6.1.4(17) for which
entities that are references only in assertions in a subprogram should
have the Proof_In global mode. We have now implemented this check.

The test filed under this ticket is currently XFAILed but will be enabled
with the work Piotr has been doing under R223-009 so there we will see the
check appearing.

* gnat2why/flow/flow-analysis.ads
* gnat2why/flow/flow-analysis.adb
(Find_Input_Only_Used_In_Assertions): add check for global inputs that
should be proof_ins.

* testsuite/gnatprove/tests/M314-012__flow_abstract_state
* testsuite/gnatprove/tests/M412-008__flow_functions_with_depends
* testsuite/gnatprove/tests/MA29-016__flow_refinement
* testsuite/gnatprove/tests/MB21-036__mainpre
* testsuite/gnatprove/tests/N813-024__flow_ineffective_constant
* testsuite/gnatprove/tests/O105-023__flow_wrong_scope_for_refined
* testsuite/gnatprove/tests/O219-034__flow_global_checks
* testsuite/gnatprove/tests/P229-022__flow_unused_with_exceptions
* testsuite/gnatprove/tests/QB29-037__flow_proof_in
* testsuite/gnatprove/tests/text_io_get_line
Update test outputs.

* testsuite/gnatprove/tests/sparksmt__names
Adjust test to not have failed checks (fix global mode) and update test
output.

Change-Id: I8f209632f05303169ba83fbdfe7ae8cbaead7fdb