R809-002 translate declaration block into a flat sequence
Lists of Ada declarations were translated into a tree of Why3 sequence
nodes; if the list was long, then the tree was deep and we crashed while
recursively traversing it.

* gnat2why/why/gnat2why-expr.adb
(Transform_Declarations_Block): translate list of Ada declarations into a
 flat sequence of Why3 nodes. Note: previously we were iterating the list
 from tail to front, now we do this front to tail.

* testsuite/gnatprove/tests/LC17-035__update_big_test/
The crash that happened in developer builds is now fixed.

* testsuite/gnatprove/tests/MA14-020__dynamic_types/test.out
Change in counterexample due to change in the order of declaration
processing. Previously the declarations in the tail were getting small
identifiers (e.g. foo__123) and the ones in the front were getting large
identifiers (e.g. foo__321); now it is the opposite.

Change-Id: I561ed6412fe90f5c0798c089ee9909ab8b3803e1