R809-003 Fix interval sloc generation in ce code
This changes the way we generates the intervals associated to the removal
of some counterexamples (in the context of if_branching). We now make
sure that the left bound is always less than the right_bound.

* gnat2why/why/gnat2why-counter_examples.adb
(Get_Interval_For_Branch): Compute left_bound and then constrain the right
bound to be higher by using Max.

Change-Id: I01be542adce7f04d4c68b971c94ed8b21e8694e6