Q123-035 strip test to the bare minimum
Looking at the CFGs generated for this test the problem happens in a call to
a null procedure, whose Global contract is not verified (since null
procedures have no body). The rest is unimportant.

* testsuite/gnatprove/tests/Q123-035__flow_null_subp_with_globals/
Minimize; update output (it still info messages, which are wrong, and lacks
some checks that we would expect).

Change-Id: I6b1b67b41c1a762489bccdbc0d0f27fa9264f6bc