minor use non-array variant of Sequence where possible
We have a call to Sequence that takes an array of items:

  function Sequence (Progs : W_Prog_Array) return W_Prog_Id;

which is meant to be used with more than 2 items (or when the number of
items is not known in advance). There is no point in calling it where the
number of items is exactly 2.

(no-tn-check)

(Unroll_Loop): change to non-array variant of Sequence.
(Comp_Decl_At_Subp_Start): same as above.

Change-Id: Ia932d23e5750c1ac61a8f97e92af0bc906a5dca2