R810-016 Mark all lemmas from SPARK lemma library as terminating
Lemmas from the SPARK lemma library are terminating procedures, but since
the library is used in a context where their body is marked SPARK_Mode=>Off,
they are considered by default as potentially terminating by GNATprove when
called from a function. This leads to spurious messages about these
functions potentially not terminating when they are called from contracts
and assertions. Mark the lemmas as terminating to avoid that issue.

* include/manual_proof.in
Adjust lines for manually proved lemmas.

* include/spark-arithmetic_lemmas.ads
* include/spark-constrained_array_lemmas.ads
* include/spark-fixed_point_arithmetic_lemmas.ads
* include/spark-floating_point_arithmetic_lemmas.ads
* include/spark-mod_arithmetic_lemmas.ads
* include/spark-unconstrained_array_lemmas.ads
Add Annotate pragma to state that all lemmas in the library unit are
terminating.

* testsuite/gnatprove/tests/P405-006__arithmetic_lemmas/test.out
Update output.

Change-Id: I3bf94975997912cdfeb78b616f2f1b79240893ab