R509-027 Handling differently constants and In parameters.
We allow an IN parameter or a constant of an owning access type to
provide read/write access to its designated object.
In SPARK, IN access-to-variable provides this access for a procedure
and not for a function call. This is because the former is considered
as a borrow operation while the former is an observe one.

* gnat2why-expr.adb
(Needs_Temporary_Ref): gives a temporary ref to IN parameters of an access
type in a function call and which need conversions.
(Compute_Param): IN parameters of an access type in a function call
are considered as IN OUT parameters.
(Transform_Expr): Use E_Symb to get correct allocator symbol.

* gnat2why-util.adb
(Is_Mutable_In_Why): constants and IN parameters of an access
type in a function call are considered mutable.

* why-atree-modules.adb
(Insert_Type_Symbols): Fix type and module of WNE_Uninit_Allocator
and WNE_Init_Allocator.

* why-gen-pointers.adb
(Declare_Access_Function): simplifying a condition.
(New_Pointer_Value_Access): Use program function in Program domain.

* why-inter.adb
(Need_Conversion): Also apply conversion on allocators.

Change-Id: I265db160fcd50c34304c6f26dd57aa010f2c1f90