R618-018 Improve check message on missing global
For checks related to a global missing from a Refined_Global or
Refined_Depends we now issue a continuation message suggesting to add this
global to the corresponding Global or Depends (or even both if they are
both present) in case it is missing from there as well.

* gnat2why/flow/flow-analysis-sanity.adb
(Check_All_Variables_Known): add continuation message for checks related
to missing globals from Refined_Global/Depends

* testsuite/gnatprove/tests/R618-018__flow_global_check_message/
* testsuite/gnatprove/tests/R618-018__flow_global/
* testsuite/gnatprove/tests/R618-018__flow_extra_global/
Add tests.

Change-Id: Iedba0d89c136e4a948fbe09bc7fb4a7b6f07a91a