remove dead code from Type_Of_Node
The CASE expression in the return statement of Type_Of_Node is not needed,
because T must be a type (this is evident from the body). Also, if it would
be a type, then Retysp would crash on its precondition.

(no-tn-check)

* gnat2why/spark/spark_util-types.ads
(Retysp): add obvious postcondition.

* gnat2why/why/gnat2why-util.ads
* gnat2why/why/gnat2why-util.adb
(Type_Of_Node): remove unnecessary CASE expression; propagate an obvious
 postcondition from Retysp.

Change-Id: Ie2657be1e372e281191df3a248c2afad00468e21