R805-001 support package "Inputs"
The Initializes can't capture objects used for the IN mode actual parameters
of generic instances or in the declaration blocks in the package body
statements. We only pick them when the (generated) Initializes contract is
null, as otherwise they would appear in the (generated) RHSs.

* gnat2why/flow/flow-control_flow_graph.adb
(Create): add vertices for the package "Inputs".

* testsuite/gnatprove/tests/O909-044__flow_wrong_gen_initializes/test.out
* testsuite/gnatprove/tests/PA17-029__flow_same_variable_name/test.out
* testsuite/gnatprove/tests/R805-001__flow_generic/test.out
Spurious checks are gone.

Change-Id: I04ee5ffdf9b11a90a0e966817cac0203a2de9d4f