R831-022 get correct is_mutable for identifiers
When transforming an identifier which is in the symbol map, we used to look
at the Ada identifier to decide if we need a dereference in Why or not. This
was fragile as the Ada and the Why entity do not need to have the same
mutable status.
Rather use information stored in the symbol map to decide if a dereference
is needed.

* gnat2why/why/gnat2why-expr.adb
(Transform_Identifier): Use the Why3 item and not the Ada node to decide
 whether we need a dereference when transforming an identifier.

Change-Id: I17b1f72746def2bf8b1e9478587ea5333fb40910