R718-010 fix detection of mixed-default-initialized records
For record components the Is_Default_Initialized, which is used to set the
CFG vertex attribute Is_Initialized, only looked at the DIC on the entire
object. But when one component has a DIC and other doesn't, the entire
object can't have a DIC, some of its components are always initialized.

* gnat2why/flow/flow_utility-initialization.adb
(Is_Default_Initialized): pick a DIC while climbing the component hierarchy.

* testsuite/gnatprove/tests/R718-010__flow_initialized_state2/test.out
Spurious warning is gone.

Change-Id: Ic67e297205a1e4e731e1ed3083e27399303be8f1