minor move Default_Initialization for Flow_Utility.Initialization
Moved from Flow_Refinement, where it was for no obvious reason (perhaps
because it needs a Flow_Scope parameter, but that seems a big deal).

Also, move Default_Initialization_Kind from Flow_Types to a dedicated
package that deals with a type initialization. The Flow_Types package (just
like the frontend's Types) is only meant for essential declarations that
depend are used everywhere and depend on each other (e.g. Node_Id that
logically belongs to sinfo.ads and Entity_Id that logically belongs to
einfo.ads).

(no-tn-check)

Change-Id: I11273542b0def883206e3b12a024931b275d7a66