R906-041 incorrect Constrained attribute in record components
When assuming default initialization of a record component for which a
default expression is supplied, we incorrectly take the value of the
'Constrained attribute of such components which have mutable discriminants
from the default expression. Since the default expression is a constant, its
'Constrained attribute is always false whereas components of an
unconstrained type with mutable discriminants are supposed to always be
unconstrained.

* gnat2why/why/gnat2why-expr.adb
(Compute_Default_Init): When taking the value of a record component from
a default expression, reset the constrained attribute to the one enforced
by the component type.

Change-Id: Ic0d9d8ea3568a1fa50184f15dec478f920881875