Q802-030 Output info messages about inlining and unrolling
When --info is set as done in the testsuite, GNATprove now outputs
additional information about the inlining of subprograms and the unrolling
of loops, when the corresponding subprogram or loop is a candidate for
inlining or unrolling respectively.

* gnat2why/spark/spark_definition.adb
(Check_Unrolled_Loop): Adapt to new signature for
 Candidate_For_Loop_Unrolling.

* gnat2why/spark/spark_util.adb
* gnat2why/spark/spark_util.ads
(Candidate_For_Loop_Unrolling): Add parameter Output_Info so that a reason
 is printed for not unrolling when possible, and a message is output for
 successful unrolling when relevant.

* gnat2why/why/gnat2why-expr-loops.adb
(Transform_Loop_Statement): Call Candidate_For_Loop_Unrolling with new
 parameter.

* testsuite/gnatprove/tests/*/test.out
Update test output.

Change-Id: Ida1172fb913e0b869cbe5415b7c88e3fd85b0059