R509-027 Dealing with private access types
* gnat2why-driver.adb
(Translate_Entity): add a case when the translated entity is mutable
and has a partial view.

* gnat2why-expr.adb
(One_Level_Access): simplification + using the representative type
in SPARK to call New_Pointer_Value_Access when dereferencing a
variable of access type.
(Transform_Comparison; Transform_Expr): using the representative type in
SPARK to compare variable of access type and determine the allocator type.

* gnat2why-util.adb
(Is_Mutable_In_Why): private types whose fullview is an access type
which are constants are mutable (Recall that we allow an IN
parameter or a constant of an owning access type to provide
read/write access to its designated object).

* why-gen-pointers.adb
(Declare_Equality_Function): use the Pred domain instead of
Term one (simplification).
(Root_Pointer_Type): extract the Directly_Designated_Type from
the entity instead of it's Etype because the information is lost
for private access types.

Change-Id: Ib2068881214978719866bd5cce456ccdb4404542