PC08-018 Default initialization for private types
Flow analysis was failing to correctly determine the default initialization
status for a type derived from a private type with a
Default_Initial_Condition. This is now fixed.

* gnat2why/flow/flow_utility-initialization.adb
(Default_Initializaton): for private types, if their full view isn't
visible from the Scope we are looking from it means that we can trust a DIC
annotation if present.

* testsuite/gnatprove/tests/PA12-001__DIC
Update test output. Wrong check is now gone.

Change-Id: I72811f1069d05c59b5711420aff95231fe70e7b9