R917-060 fix Get_Flow_Scope for separate bodies
When calling Subprogram_Refinement_Is_Visible on subprogram with a separate
body we wrongly returned the scope of where that subprogram is declared, not
where its body stub is.

This could result in extra trimming (or missing trimming, I am not sure...)
of user-written Depends based on Refined_Global or Refined_Depends (as would
be exhibited by an assertion added to Get_Depends when executed on
M729-011__flow_body_stubs).

Now fixed.

* gnat2why/flow/flow_refinement.adb
(Get_Flow_Scope): restart query when jumping from a subunit to its parent
 unit.

* gnat2why/flow/flow_utility.adb
(Get_Depends): add assertion that would crash without the above fix.

Change-Id: I43523aece6b102b30dd58cb6a1e006294f55ab48