R321-005 Add check for objects missing from Initializes
GNATprove was not enforcing SPARK RM 7.1.5(9) completely. We now emit
checks for all the initialized objects that are missing from the
Inizializes contract, including default initialized objects and objects
initialized by the package body statements.
We also move some code that was previously in Find_Ineffective_Statement
and that was detecting objects missing from an Initializes. That was a
warning but now has been upgraded to a check, consistently with the SPARK
RM rule 7.1.5(9).

* gnat2why/flow/flow-analysis.adb
(Find_Ineffective_Statements): remove code that was checking for an object
missing from an Initializes.
(Check_Initializes_Contract): add code to check for objects and/or clauses
missing from an Initializes. Enable the checks also for a null Initializes.

* testsuite/gnatprove/tests/L118-012__bool
* testsuite/gnatprove/tests/MA15-019__flow_contract_check_initializes
* testsuite/gnatprove/tests/N701-018__flow_raise_statements
* testsuite/gnatprove/tests/O318-015__init
* testsuite/gnatprove/tests/P819-010__flow_locals
Update test output. Warning is now a check.

* testsuite/gnatprove/tests/M314-014__flow_packages
* testsuite/gnatprove/tests/NC18-019__flow_nested_package_bodies
* testsuite/gnatprove/tests/P330-009__useless_msg_on_task_ob_init
* testsuite/gnatprove/tests/TU__initializes_illegal
Update test output. Warning is now a check, wrong warning disappears and/or
new check on object to be listed in the Initializes contract.

* testsuite/gnatprove/tests/M618-009__second
* testsuite/gnatprove/tests/MC05-017__flow_initializes_issues
* testsuite/gnatprove/tests/O429-046__flow_tasking_cfg
* testsuite/gnatprove/tests/Q929-028__flow_spurious_error
* testsuite/gnatprove/tests/QA24-010__flow_race
* testsuite/gnatprove/tests/QA25-021__flow_task_init
* testsuite/gnatprove/tests/TU__refined_global_legal
Update test output. New checks for objects to be listed in the Initializes
contract.

* testsuite/gnatprove/tests/Q830-018__type_invariant
Enable test. Add test output.

* testsuite/gnatprove/tests/R321-005__flow_initializes
* testsuite/gnatprove/tests/R321-005__flow_initializes_nested_packages
* testsuite/gnatprove/tests/R321-005__flow_initializes_private_child
Add tests

Change-Id: I2fed093e1d7148b16e29ec1308846310b894ebb2
Signed-off-by: Elisa Barboni <elisa.barboni@altran.com>