P831-036 disable simplification for constants in URA
When untangling non-local constants that appear on the RHS of an assignment
statement we were "simplifying" them, i.e. pretending that each of the
assigned component depends on *all* components of the constant (as retrieved
by Get_Variables_Wrapper).

This was OK, because non-local constant is seen either as a single blob or
as an empty set (depending whether it has variable input or not). As it is
both global and immutable, we don't need to track the flow of information
via its individual components (like we do for local constants and local &
global variables). However, it required us to know which constants are local
and carrying this information was obfuscating the code (*).

Now we create a map where each component on the LHS depends on the
corresponding component of the RHS; if the RHS is a "pure" constant (i.e.
has no variable input), then the component depends on nothing.

This works both for local and global constants.

(*) Actually, we could decide which constants are local by the Flow_Scope
parameter of URA...

* gnat2why/flow/flow_utility.adb
(Untangle_Record_Assignment): disable "simplification" for non-local
 constants.

Change-Id: I6584138b212c7ea47fbd9a4e8b3ea51633db4324