OA19-025 Discrete type counterexample printing
This changes the way counterexample are printed: if the counterexample is
close enough to its type bound then it is expressed as a function of this
bound. With counterexample,  (V : Integer = -2147483648), it prints
(V = Integer'First). This optimization in the printing is done only for
the expected type of the counterexample:
(V : Short_Short_Integer =  -2147483648) is printed (V = -2147483648)

* gnat2why/why/ce_pretty_printing.ads
(Print_Discrete): Generic function that takes the value of a
counterexample and the discrete type of the associated node and returns
the new printing. Bound_Type is used to limit this simplification to
types of range greater than 10 and Bound_Value to limit the range of
values printed with the new simplification.

* gnat2why/why/ce_pretty_printing.adb
(Beautiful_Source_Name): Prints the source name associated to a type.
Source_Name uses the casing used by the user in the definition to pretty
print an element: for standard types, it is not possible to do that so we
just print them with an upper case letter at the beginning of each word.
(Print_Discrete): Compute the differences between the value and the bounds
of the types to chose whether to alter the printing or not.

* gnat2why/spark/spark_util.ads
(Is_Standard_Entity): Popular hack to know if an entity is defined during
stdlib pass.

* gnat2why/why/gnat2why-counter_examples.adb
(Refine_Aux): Use Print_Discrete for discrete types.

* testsuite/gnatprove/*/*/test.out
Update outputs

Change-Id: Ifc393d8f4b242721863e81d3896aa4278c6802b4