R920-019 raise violation on aggregates when components are not in SPARK
Move the code checking whether a component is in SPARK used for checking
selected components to the marking of the component entity so that it can
be used when marking an aggregate.

* gnat2why/spark/spark_definition.adb
(Mark): Do not check if the component is visible when marking a selected
 component as this will be done as part of marking the selector name.
(Mark_Identifier_Or_Expanded_Name): Raise a violation on a component or
 discriminant which is not visible in SPARK.

* gnat2why/spark/spark_util.adb
(Component_Is_Visible_In_SPARK): Discriminants may not be in SPARK if they
 are completely hidden by a private type declaration.

* tests/P801-002__invariant_on_tagged_component/test.out
* tests/P801-002__invariant_on_tagged_types/test.out
* tests/P829-014__invariant/test.out
* tests/TU__type_invariant_tagged__legal/test.out
Violation are raised on components and not on selection.

Change-Id: Ibc2b027453f618830c0a4040718346ed2a35803a