R509-027 Fix and differently handling of the null exclusion
The null exclusion check is now not related only to allocation
but to assignment to a pointer in general.

* gnat2why-expr.adb
(Transform_Expr): remove the call to the vc_null_exclusion.

* why-atree-modules.adb
(Insert_Type_Symbols) add a symbol related to assigning to a pointer

* why-gen-expr.adb
(Insert_Pointer_Conversion) add a call to the vc_null_exclusion
when needed.

* why-gen-names.adb
(To_String): add WNE_Assign_Null_Check symbol

* why-gen-pointers.adb
(Declare_Allocation_Function) remove the check on the vc_null_exclusion
from the uninitialized call.
(Declare_Access_Function) remove a program function related to the
vc_null_exclusion and add the call to check the null exclusion on
assignments.

Change-Id: Ie020e01b8b11d01e0393493bf8f8c45a6fbb13a7