R805-001 map generic IN formals when seen from the outside
Generic IN formal parameters can only appear inside generic instances;
outside of those instances, they should be mapped to objects used in their
corresponding generic actual parameter expressions.

This patch does this for user-written Global, for both user-written and
generated Initializes, and for Get_Variables too. This is the minimum
required to get this idea to work on existing tests. It also does this for
Depends, which appears to make sense, but doesn't affect any of our tests.

Note: this patch does NOT implement this idea for generated globals,
because (in general) functions used in the generic actual parameter
expressions might themselves have no globals yet (e.g. when they are
declared in other compilation units); that is for phase 1. This is even
harder in phase 2, because deciding when to map actuals depend on visibility
and we don't have it yet there.

In particular, some tests got explicit Global annotations to benefit from
this patch (and to demonstrate that it actually works); I see no point in
keeping those tests broken, as that doesn't guard us against regressions.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Object_Declaration): ignore all generic actuals.

(Do_Package_Body_Or_Stub): map the RHS of Initializes contract.

* gnat2why/flow/flow_utility.adb
(Get_Global): map Global contract & reshuffle the code to allow this.
(Get_Depends): map Depends contract.

(Map_Generic_In_Formals): new routine that implements mapping.

* testsuite/gnatprove/tests/N423-024__generic_inst/context.adb
Adapt test; output is now as expected.

* testsuite/gnatprove/tests/O206-030__generic/test.out
* testsuite/gnatprove/tests/O917-119__crazyflie/test.out
* testsuite/gnatprove/tests/PC06-008__flow_generic_actual/test.out
* testsuite/gnatprove/tests/Q217-004__flow_generic_in_parameters/test.out
* testsuite/gnatprove/tests/R122-008__flow_iterate_generic_parameters/...
Dubious info and check messages are gone. They overlapped with legitimate
errors "actual for formal object of mode in cannot depend on variable
input", so this is an improvement. This comes from ignoring generic actuals
in Do_Object_Declaration.

* testsuite/gnatprove/tests/P518-021__frame_for_max/frame_condition.adb
Add Global contract to keep the output as it was (somehow it worked in the
old setting).

* testsuite/gnatprove/tests/R805-001__flow_generic_map/test.out
* testsuite/gnatprove/tests/R805-001__flow_generic/test.out
* testsuite/gnatprove/tests/Q927-020__flow_generic/test.out
Primary improvements; spurious checks are gone.

Change-Id: Ic8d7fc72fbf4043286661b9a7cb42bbc0f0b7ae4