R926-017 keep Depends clauses with empty RHS when trimming
When trimming the RHS of a down-projected, user-written Depends contract by
intersecting it with the Refined_Global, we shall keep the clause even if
the Refined_Global is empty (which happens especially in phase 1 before the
Refined_Global has been generated).

The reason for discarding the dependency clause is a mistake while removing
an old, semantically-broken error recovery in R426-033.

* gnat2why/flow/flow_utility.adb
(Get_Depends): keep the clause while trimming, even if its RHS is empty.

* testsuite/gnatprove/tests/R926-017__depends/
Enable test; add output.

Change-Id: Ie777bfb2c08f4662de4b4678c1a92aa5c8dd8da9